// src/types/node.ts
export type NodeDataType = ResponseLocationType | string | number | boolean | UploadedFileType;

export type NodeDataTypes = NodeDataType[];

export type NodePropertiesValues = {
  id: string;
  node_id?: string;
  name: string[];
  default_image: UploadFile[];
  project_type_property_id?: string;
  project_type_property_type?: string;
  nodes_data?: NodeDataTypes;
  nodeTypeProperty: {
    name: string;
    id: string;
    default_image: boolean;
    default_property: boolean;
    multiple_type: boolean;
    enums_data: IEnumProperty[];
  };
  project_type_id: string;
  user_id: string;
  project_id: string;
  created_at: string;
  updated_at: string;
  createdEdges?: ICreateEdge[];
};
// ...
export interface ICreateEdge {
  id: string;
  label: string;
  source_id: string;
  target_id: string;
  project_edge_type_id: string;
}